Add a black box invariant that executes an external function

+++ src/invariantgraph/violationInvariantNodes/boolLinRelNode.cpp
@@ -158,7 +158,7 @@ void BoolLinRelNode::updateState() {
       setState(InvariantNodeState::SUBSUMED);
       return;
     }
-    for (long& coeff : _coeffs) {
+    for (Int& coeff : _coeffs) {
       coeff /= c;
     }
     _rhs /= c;
